+++ src/api/polls.ts
@@ -10,7 +10,7 @@ export interface PollOption {
 
 export interface Poll {
   id: string,
-  expires_at: string,
+  expires_at: string | null,
   expired: boolean,
   multiple: boolean,
   votes_count: number,