+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Do not allow replying with "Subscribers only" visibility to public posts.
+
 ## [4.12.0] - 2025-10-26
 
 ### Changed