+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [4.13.0] - 2025-11-13
+
 ### Added
 
 - Add alt text tooltips to attached audio and video.